Documentation: Update the code of conduct
This change reiterates that the coreboot leadership may revoke a user's
privileges. This does not change anything, as it already says "The
community organizers may take any action they deem appropriate, up to
and including a temporary ban or permanent expulsion from the community
without warning".

Also add a note that the discussions are private. If someone wants to
make the issue public, the coreboot leadership can't stop them, but the
board believes in handling these issues privately.

Finally, add a note that if there's an issue with someone on the
arbitration board, issues may be taken directly to the leadership board.

Change-Id: I5e2010a16f31f892bd1761b56b96ea773877dea0
Signed-off-by: Martin Roth <gaumless@gmail.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/80619
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Eric Lai <ericllai@google.com>
Reviewed-by: David Hendricks <david.hendricks@gmail.com>